rename file name instead of changing routers
