+++ docs/src/app/components/pages/get-started/Prerequisites.jsx
